Rename consts.ts to enums.ts (that's what's in there now!)
