+++ packages/math-input/src/components/expression-keypad.tsx
@@ -6,7 +6,7 @@ import {StyleSheet} from "aphrodite";
 import * as React from "react";
 import {connect} from "react-redux";
 
-import {BorderStyles} from "../consts";
+import {BorderStyles} from "../enums";
 import KeyConfigs from "../data/key-configs";
 import {View} from "../fake-react-native-web/index";
