+++ packages/math-input/src/components/fraction-keypad.tsx
@@ -6,8 +6,8 @@
 import * as React from "react";
 import {connect} from "react-redux";
 
-import {BorderStyles} from "../consts";
 import KeyConfigs from "../data/key-configs";
+import {BorderStyles} from "../enums";
 import {View} from "../fake-react-native-web/index";
 
 import {CursorContext} from "./input/cursor-contexts";
