+++ packages/math-input/src/components/keypad-button.tsx
@@ -6,7 +6,7 @@ import {StyleSheet, css} from "aphrodite";
 import * as React from "react";
 import {connect} from "react-redux";
 
-import {BorderDirection, BorderStyles, KeyType} from "../consts";
+import {BorderDirection, BorderStyles, KeyType} from "../enums";
 import {View} from "../fake-react-native-web/index";
 
 import {
