+++ packages/math-input/src/components/keypad-container.tsx
@@ -2,7 +2,7 @@ import {StyleSheet} from "aphrodite";
 import * as React from "react";
 import {connect} from "react-redux";
 
-import {LayoutMode, KeypadType} from "../consts";
+import {LayoutMode, KeypadType} from "../enums";
 import {View} from "../fake-react-native-web/index";
 import {setPageSize} from "../store/actions";
