+++ packages/math-input/src/components/multi-symbol-grid.tsx
@@ -6,7 +6,7 @@
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
 
-import {IconType} from "../consts";
+import {IconType} from "../enums";
 import {View} from "../fake-react-native-web/index";
 import {IconConfig} from "../types";
