+++ packages/math-input/src/components/multi-symbol-popover.tsx
@@ -5,7 +5,7 @@
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
 
-import {BorderStyles} from "../consts";
+import {BorderStyles} from "../enums";
 import {View} from "../fake-react-native-web/index";
 import {KeyConfig} from "../types";
