+++ packages/math-input/src/components/navigation-pad.tsx
@@ -6,7 +6,7 @@
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
 
-import {BorderStyles} from "../consts";
+import {BorderStyles} from "../enums";
 import KeyConfigs from "../data/key-configs";
 import {View} from "../fake-react-native-web/index";
