+++ packages/math-input/src/enums.ts
